fix sample endpoints
