+++ genotype_api/service/match_genotype_service/utils.py
@@ -1,5 +1,8 @@
 """Utils functions for the match genotype service."""
 
+from collections import Counter
+
+from genotype_api.constants import CUTOFS, SEXES
 from genotype_api.database.models import Genotype
 
 
@@ -12,3 +15,40 @@ def compare_genotypes(genotype_1: Genotype, genotype_2: Genotype) -> tuple[str, 
         return genotype_1.rsnumber, "match"
     else:
         return genotype_1.rsnumber, "mismatch"
+
+
+def check_sex(sample_sex, genotype_analysis, sequence_analysis):
+    """Check if any source disagrees on the sex"""
+    if not sample_sex or genotype_analysis.sex == SEXES.UNKNOWN:
+        return "fail"
+    sexes = {genotype_analysis.sex, sequence_analysis.sex, sample_sex}
+    if {SEXES.MALE, SEXES.FEMALE}.issubset(sexes):
+        return "fail"
+    return "pass"
+
+
+def check_snps(genotype_analysis, sequence_analysis):
+    genotype_pairs = zip(genotype_analysis.genotypes, sequence_analysis.genotypes)
+    results = dict(
+        compare_genotypes(genotype_1, genotype_2) for genotype_1, genotype_2 in genotype_pairs
+    )
+    count = Counter([val for key, val in results.items()])
+    unknown = count.get("unknown", 0)
+    matches = count.get("match", 0)
+    mismatches = count.get("mismatch", 0)
+    snps = (
+        "pass"
+        if all([matches >= CUTOFS.get("min_matches") and mismatches <= CUTOFS.get("max_mismatch")])
+        else "fail"
+    )
+    nocalls = "pass" if unknown <= CUTOFS.get("max_nocalls") else "fail"
+    failed_snps = [key for key, val in results.items() if val == "mismatch"]
+
+    return {
+        "unknown": unknown,
+        "matches": matches,
+        "mismatches": mismatches,
+        "snps": snps,
+        "nocalls": nocalls,
+        "failed_snps": failed_snps,
+    }
